# src/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/health/live": {
      "get": {
        "operationId": "HealthController_liveness",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/health/ready": {
      "get": {
        "operationId": "HealthController_readiness",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/status": {
      "get": {
        "operationId": "StatusController_getStatus",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "summary": "Worker login with email and password",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login success, returns opaque bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation error - problem-details"
          },
          "401": {
            "description": "Invalid credentials (generic, no existence leak)"
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/me": {
      "get": {
        "operationId": "MeController_me",
        "summary": "Get current authenticated user",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Current user"
          },
          "401": {
            "description": "Unauthorized - invalid or missing token"
          }
        },
        "tags": [
          "auth"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "summary": "Logout current session",
        "parameters": [],
        "responses": {
          "204": {
            "description": "Logged out, session revoked"
          },
          "401": {
            "description": "Unauthorized - invalid or missing token"
          }
        },
        "tags": [
          "auth"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "Buildflow API",
    "description": "Buildflow API \u2014 Worker self-claim slice",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "opaque",
        "type": "http"
      }
    },
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "worker1@example.com",
            "description": "Worker email"
          },
          "password": {
            "type": "string",
            "example": "Password123!",
            "description": "Account password"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "LoginResponseUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "clx123"
          },
          "email": {
            "type": "string",
            "example": "worker1@example.com"
          },
          "role": {
            "type": "string",
            "example": "WORKER",
            "enum": [
              "WORKER",
              "ADMIN",
              "COORDINATOR"
            ]
          },
          "status": {
            "type": "string",
            "example": "ACTIVE",
            "enum": [
              "ACTIVE",
              "LOCKED",
              "DISABLED"
            ]
          }
        },
        "required": [
          "id",
          "email",
          "role",
          "status"
        ]
      },
      "LoginResponseDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "example": "a1b2c3..."
          },
          "user": {
            "$ref": "#/components/schemas/LoginResponseUserDto"
          }
        },
        "required": [
          "token",
          "user"
        ]
      }
    }
  }
}